rename repro-templates -> sandbox-templates
