+++ scripts/next-repro-generators/utils/template.ts
@@ -2,7 +2,7 @@ import { render } from 'ejs';
 import { readFile } from 'fs-extra';
 import { format } from 'prettier';
 import type { GeneratorConfig } from './types';
-import { allTemplates as reproTemplates } from '../../../code/lib/cli/src/repro-templates';
+import { allTemplates as reproTemplates } from '../../../code/lib/cli/src/sandbox-templates';
 
 export async function renderTemplate(templatePath: string, templateData: Record<string, any>) {
   const template = await readFile(templatePath, 'utf8');
